feat: verification implementation
# Motivation

This commit adds the implementation of the verification command. The
interface has been specified in commit
d6a9fd5c6a8d99ac36cca361fe65af1dad594f69, but the interface was expanded
in commit 7cb6c6e33d19bbeddcc641721c2807157856f933.

# Implementation

In summary, the verification will happen like this:

1. Read the next message from the consumer group
2. If the message is not owned by the web archive, return to step 1.
3. If we get a read timeout, sleep for 1 minute and return to step 1.
4. Create a Teams message and send it to the relevant channel
5. Return to step 1.

# Considerations

If the verification service fails between steps 1 and 4, a message could
be read from the consumer group, but will not be sent to the Teams
channel successfully. This means that the message will be lost.

# Future work

If unexpected errors occur no message will be sent to Teams and it will
be difficult to know that the service has failed. This could be remedied
by always sending a message to Teams whenever an error is detected.