feat: add common arguments
# Motivation

The main goal of this commit is to avoid duplicating the same arguments
for all subcommands.

# Implementation

This commit adds `--teams-webhook-notification-url` as a top-level
argument and also adds parsing of this flag and the existing
`--kafka-endpoints` argument. This is done automatically by
`StringSliceVar` and `StringVar`.

Both will be used by the `verification` and `transfer` implementations
in later commits.