feat: avoid resending same messages
# Motivation

To avoid sending the same messages again and again if the service is
restarted, we now take into account the previous state of the `kafka`
topic.

This ensures we will get a lot fewer false-positive notifications when
notifications are enabled.

# Implementation

In short, this is what happens:

* All messages between (inclusive) the first and last offsets are read.
* The messages that do not belong to `nettarkivet` are filtered away
* The remaining messages `Path` value are put in a cache
* The cache is consulted before sending a message to the transfer topic
   - If the `Path` evaluated is already in the cache, it is not sent
   - Otherwise, a `kafka` message is sent to the transfer topic, and a
     new entry is added to the cache

In addition, now the program will not exit after sending the messages,
but rather run a forever loop every 1 minute.

# Quirks

## Cache configuration

The configuration of the cache might be tuned to avoid memory issues.
Each of the entries in the cache is stored there for 24 hours, but this
might need to be evaluated if the amount of data we send increases.

## Valid offsets

Currently, all offsets between the first and last offset are read,
including the first and lasts. But sometimes certain offsets might be
deleted, these are currently ignored, but could potentially be a source
of bugs in the future.

## `kafka.FirstOffset` and `kafka.LastOffset`

Neither of these constants are used as these offsets are not
"evaluated", i.e. turned from `-1` and `-2` to the actual offsets, until
after a message has been read. This means that it is difficult to use
these constants to determine when to stop reading messages.