fix: send `Teams` error message on system failure
# Motivation

If something unexpected happens, like an unhandled error, a message will
be sent to `Teams` notifying the users that something is wrong. This
will reduce the cognitive load of the users since they don't have to
check the logs to see if something is wrong.

# Future work

The errors could be unwrapped for a nicer user experience, but at this
point, it seems like premature optimization.