fix: use `UUID` in identifier
# Motivation

Commit 60d388416df2c3fc89cbe338226cf2f6e27096eb mistakenly assumed that
`URN`s are unchangeable, but this is not the case. In certain
situations, there might be a reupload of the same `URN`, for example
when something was uploaded with the wrong name.

The true unchangeable reference to the object is the `identifier`, which
with the current implementation is not going to be unique on a reupload.

Ideally the `identifier` should not be provided by `hermetic`, but be
returned by `Digital Preservation System` (DPS), but unfortunately this
is not the case.

# Workaround

To avoid this problem we simply add a `UUID` to the end of the
`identifier` and unlink the `URN` from the `identifier`. The UUID
version is version 4.

Theoretically, there is a chance that the `UUID` is not unique, but the
likelihood is, according to
[Wikipedia](https://en.wikipedia.org/wiki/Universally_unique_identifier#Collisions),
this:

```
For example, the number of random version-4 UUIDs which need to be
generated in order to have a 50% probability of at least one collision
is 2.71 quintillion
```